app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.5
python-multipart==0.0.19
jinja2==3.1.4
aiofiles==24.1.0
itsdangerous==2.2.0
